About
The purpose of this project is to develop an application based on the Android system that can optimize travel routes from selected locations by incorporating information on air pollution data. The air pollution data are based upon the data measured and made available by the AirCasting Project (http://www.aircasting.org). To access this data, the application sends a request to AirCasting's data server. The server parses the relevant data and returns them to the mobile application. The application integrates and calculates the processed data, and finally calculates and draws different routes for visual display. In the process of implementing this project, there are many challenges related to application development. The report also details the development process undertaken throughout this work.